add implimentation of pricing Methods
add implimentation of pricing Methods
check tests

+++ kata-supermarket-pricing/src/main/java/service/Pricing.java
@@ -4,7 +4,7 @@ import model.Item;
 
 public abstract class Pricing {
 
-	    public abstract double calculatePrice(Item item, double numberBought);
+	    public abstract double calculatePrice(Item item, Integer numberBought);
 
 	}
